+++ philosophers/libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: amouly <amouly@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/03 17:33:32 by amouly            #+#    #+#             */
-/*   Updated: 2023/02/13 15:25:55 by amouly           ###   ########.fr       */
+/*   Updated: 2023/02/13 15:58:59 by amouly           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-long long int	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
 	int	i;
 	int	neg;
-	long long num;
+	int num;
 
 	i = 0;
 	num = 0;
